// src/opamp/generated/anyvalue_pb.d.ts
export declare type AnyValue = Message<"opamp.proto.AnyValue"> & {
  /**
   * The value is one of the listed fields. It is valid for all values to be unspecified
   * in which case this AnyValue is considered to be "null".
   *
   * @generated from oneof opamp.proto.AnyValue.value
   */
  value: {
    /**
     * @generated from field: string string_value = 1;
     */
    value: string;
    case: "stringValue";
  } | {
    /**
     * @generated from field: bool bool_value = 2;
     */
    value: boolean;
    case: "boolValue";
  } | {
    /**
     * @generated from field: int64 int_value = 3;
     */
    value: bigint;
    case: "intValue";
  } | {
    /**
     * @generated from field: double double_value = 4;
     */
    value: number;
    case: "doubleValue";
  } | {
    /**
     * @generated from field: opamp.proto.ArrayValue array_value = 5;
     */
    value: ArrayValue;
    case: "arrayValue";
  } | {
    /**
     * @generated from field: opamp.proto.KeyValueList kvlist_value = 6;
     */
    value: KeyValueList;
    case: "kvlistValue";
  } | {
    /**
     * @generated from field: bytes bytes_value = 7;
     */
    value: Uint8Array;
    case: "bytesValue";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message opamp.proto.AnyValue.
 * Use `create(AnyValueSchema)` to create a new message.
 */
export declare const AnyValueSchema: GenMessage<AnyValue>;

/**
 * ArrayValue is a list of AnyValue messages. We need ArrayValue as a message
 * since oneof in AnyValue does not allow repeated fields.
 *
 * @generated from message opamp.proto.ArrayValue
 */
export declare type ArrayValue = Message<"opamp.proto.ArrayValue"> & {
  /**
   * Array of values. The array may be empty (contain 0 elements).
   *
   * @generated from field: repeated opamp.proto.AnyValue values = 1;
   */
  values: AnyValue[];
};
// ...
export declare type KeyValueList = Message<"opamp.proto.KeyValueList"> & {
  /**
   * A collection of key/value pairs of key-value pairs. The list may be empty (may
   * contain 0 elements).
   *
   * @generated from field: repeated opamp.proto.KeyValue values = 1;
   */
  values: KeyValue[];
};

/**
 * Describes the message opamp.proto.KeyValueList.
 * Use `create(KeyValueListSchema)` to create a new message.
 */
export declare const KeyValueListSchema: GenMessage<KeyValueList>;

/**
 * KeyValue is a key-value pair that is used to store Span attributes, Link
 * attributes, etc.
 *
 * @generated from message opamp.proto.KeyValue
 */
export declare type KeyValue = Message<"opamp.proto.KeyValue"> & {
  /**
   * @generated from field: string key = 1;
   */
  key: string;

  /**
   * @generated from field: opamp.proto.AnyValue value = 2;
   */
  value?: AnyValue;
};

/**
 * Describes the message opamp.proto.KeyValue.
 * Use `create(KeyValueSchema)` to create a new message.
 */
export declare const KeyValueSchema: GenMessage<KeyValue>;