// src/opamp/generated/opamp_pb.d.ts
export declare type AgentToServer = Message<"opamp.proto.AgentToServer"> & {
  /**
   * Globally unique identifier of the running instance of the Agent. SHOULD remain
   * unchanged for the lifetime of the Agent process.
   * MUST be 16 bytes long and SHOULD be generated using the UUID v7 spec.
   *
   * @generated from field: bytes instance_uid = 1;
   */
  instanceUid: Uint8Array;

  /**
   * The sequence number is incremented by 1 for every AgentToServer sent
   * by the Agent. This allows the Server to detect that it missed a message when
   * it notices that the sequence_num is not exactly by 1 greater than the previously
   * received one.
   *
   * @generated from field: uint64 sequence_num = 2;
   */
  sequenceNum: bigint;

  /**
   * Data that describes the Agent, its type, where it runs, etc.
   * May be omitted if nothing changed since last AgentToServer message.
   *
   * @generated from field: opamp.proto.AgentDescription agent_description = 3;
   */
  agentDescription?: AgentDescription;

  /**
   * Bitmask of flags defined by AgentCapabilities enum.
   * All bits that are not defined in AgentCapabilities enum MUST be set to 0 by
   * the Agent. This allows extending the protocol and the AgentCapabilities enum
   * in the future such that old Agents automatically report that they don't
   * support the new capability.
   * This field MUST be always set.
   *
   * @generated from field: uint64 capabilities = 4;
   */
  capabilities: bigint;

  /**
   * The current health of the Agent and sub-components. The top-level ComponentHealth represents
   * the health of the Agent overall. May be omitted if nothing changed since last AgentToServer
   * message.
   * Status: [Beta]
   *
   * @generated from field: opamp.proto.ComponentHealth health = 5;
   */
  health?: ComponentHealth;

  /**
   * The current effective configuration of the Agent. The effective configuration is
   * the one that is currently used by the Agent. The effective configuration may be
   * different from the remote configuration received from the Server earlier, e.g.
   * because the Agent uses a local configuration instead (or in addition).
   *
   * This field SHOULD be unset if the effective config is unchanged since the last
   * AgentToServer message.
   *
   * @generated from field: opamp.proto.EffectiveConfig effective_config = 6;
   */
  effectiveConfig?: EffectiveConfig;

  /**
   * The status of the remote config that was previously received from the Server.
   * This field SHOULD be unset if the remote config status is unchanged since the
   * last AgentToServer message.
   *
   * @generated from field: opamp.proto.RemoteConfigStatus remote_config_status = 7;
   */
  remoteConfigStatus?: RemoteConfigStatus;

  /**
   * The list of the Agent packages, including package statuses. This field SHOULD be
   * unset if this information is unchanged since the last AgentToServer message for
   * this Agent was sent in the stream.
   * Status: [Beta]
   *
   * @generated from field: opamp.proto.PackageStatuses package_statuses = 8;
   */
  packageStatuses?: PackageStatuses;

  /**
   * AgentDisconnect MUST be set in the last AgentToServer message sent from the
   * Agent to the Server.
   *
   * @generated from field: opamp.proto.AgentDisconnect agent_disconnect = 9;
   */
  agentDisconnect?: AgentDisconnect;

  /**
   * Bit flags as defined by AgentToServerFlags bit masks.
   *
   * @generated from field: uint64 flags = 10;
   */
  flags: bigint;

  /**
   * A request to create connection settings. This field is set for flows where
   * the Agent initiates the creation of connection settings.
   * Status: [Development]
   *
   * @generated from field: opamp.proto.ConnectionSettingsRequest connection_settings_request = 11;
   */
  connectionSettingsRequest?: ConnectionSettingsRequest;

  /**
   * A message indicating custom capabilities supported by the Agent.
   * Status: [Development]
   *
   * @generated from field: opamp.proto.CustomCapabilities custom_capabilities = 12;
   */
  customCapabilities?: CustomCapabilities;

  /**
   * A custom message sent from an Agent to the Server.
   * Status: [Development]
   *
   * @generated from field: opamp.proto.CustomMessage custom_message = 13;
   */
  customMessage?: CustomMessage;
};

/**
 * Describes the message opamp.proto.AgentToServer.
 * Use `create(AgentToServerSchema)` to create a new message.
 */
export declare const AgentToServerSchema: GenMessage<AgentToServer>;
// ...
export declare type AgentDisconnect = Message<"opamp.proto.AgentDisconnect"> & {
};
// ...
export declare type ConnectionSettingsRequest = Message<"opamp.proto.ConnectionSettingsRequest"> & {
  /**
   * Request for OpAMP connection settings. If this field is unset
   * then the ConnectionSettingsRequest message is empty and is not actionable
   * for the Server.
   *
   * @generated from field: opamp.proto.OpAMPConnectionSettingsRequest opamp = 1;
   */
  opamp?: OpAMPConnectionSettingsRequest;
};
// ...
export declare type AgentDescription = Message<"opamp.proto.AgentDescription"> & {
  /**
   * Attributes that identify the Agent.
   * Keys/values are according to OpenTelemetry semantic conventions, see:
   * https://github.com/open-telemetry/opentelemetry-specification/tree/main/specification/resource/semantic_conventions
   *
   * For standalone running Agents (such as OpenTelemetry Collector) the following
   * attributes SHOULD be specified:
   * - service.name should be set to a reverse FQDN that uniquely identifies the
   *   Agent type, e.g. "io.opentelemetry.collector"
   * - service.namespace if it is used in the environment where the Agent runs.
   * - service.version should be set to version number of the Agent build.
   * - service.instance.id should be set. It may be set equal to the Agent's
   *   instance uid (equal to ServerToAgent.instance_uid field) or any other value
   *   that uniquely identifies the Agent in combination with other attributes.
   * - any other attributes that are necessary for uniquely identifying the Agent's
   *   own telemetry.
   *
   * The Agent SHOULD also include these attributes in the Resource of its own
   * telemetry. The combination of identifying attributes SHOULD be sufficient to
   * uniquely identify the Agent's own telemetry in the destination system to which
   * the Agent sends its own telemetry.
   *
   * @generated from field: repeated opamp.proto.KeyValue identifying_attributes = 1;
   */
  identifyingAttributes: KeyValue[];

  /**
   * Attributes that do not necessarily identify the Agent but help describe
   * where it runs.
   * The following attributes SHOULD be included:
   * - os.type, os.version - to describe where the Agent runs.
   * - host.* to describe the host the Agent runs on.
   * - cloud.* to describe the cloud where the host is located.
   * - any other relevant Resource attributes that describe this Agent and the
   *   environment it runs in.
   * - any user-defined attributes that the end user would like to associate
   *   with this Agent.
   *
   * @generated from field: repeated opamp.proto.KeyValue non_identifying_attributes = 2;
   */
  nonIdentifyingAttributes: KeyValue[];
};

/**
 * Describes the message opamp.proto.AgentDescription.
 * Use `create(AgentDescriptionSchema)` to create a new message.
 */
export declare const AgentDescriptionSchema: GenMessage<AgentDescription>;

/**
 * The health of the Agent and sub-components
 * Status: [Beta]
 *
 * @generated from message opamp.proto.ComponentHealth
 */
export declare type ComponentHealth = Message<"opamp.proto.ComponentHealth"> & {
  /**
   * Set to true if the component is up and healthy.
   *
   * @generated from field: bool healthy = 1;
   */
  healthy: boolean;

  /**
   * Timestamp since the component is up, i.e. when the component was started.
   * Value is UNIX Epoch time in nanoseconds since 00:00:00 UTC on 1 January 1970.
   * If the component is not running MUST be set to 0.
   *
   * @generated from field: fixed64 start_time_unix_nano = 2;
   */
  startTimeUnixNano: bigint;

  /**
   * Human-readable error message if the component is in erroneous state. SHOULD be set
   * when healthy==false.
   *
   * @generated from field: string last_error = 3;
   */
  lastError: string;

  /**
   * Component status represented as a string. The status values are defined by agent-specific
   * semantics and not at the protocol level.
   *
   * @generated from field: string status = 4;
   */
  status: string;

  /**
   * The time when the component status was observed. Value is UNIX Epoch time in
   * nanoseconds since 00:00:00 UTC on 1 January 1970.
   *
   * @generated from field: fixed64 status_time_unix_nano = 5;
   */
  statusTimeUnixNano: bigint;

  /**
   * A map to store more granular, sub-component health. It can nest as deeply as needed to
   * describe the underlying system.
   *
   * @generated from field: map<string, opamp.proto.ComponentHealth> component_health_map = 6;
   */
  componentHealthMap: { [key: string]: ComponentHealth };
};

/**
 * Describes the message opamp.proto.ComponentHealth.
 * Use `create(ComponentHealthSchema)` to create a new message.
 */
export declare const ComponentHealthSchema: GenMessage<ComponentHealth>;

/**
 * @generated from message opamp.proto.EffectiveConfig
 */
export declare type EffectiveConfig = Message<"opamp.proto.EffectiveConfig"> & {
  /**
   * The effective config of the Agent.
   *
   * @generated from field: opamp.proto.AgentConfigMap config_map = 1;
   */
  configMap?: AgentConfigMap;
};

/**
 * Describes the message opamp.proto.EffectiveConfig.
 * Use `create(EffectiveConfigSchema)` to create a new message.
 */
export declare const EffectiveConfigSchema: GenMessage<EffectiveConfig>;

/**
 * @generated from message opamp.proto.RemoteConfigStatus
 */
export declare type RemoteConfigStatus = Message<"opamp.proto.RemoteConfigStatus"> & {
  /**
   * The hash of the remote config that was last received by this Agent in the
   * AgentRemoteConfig.config_hash field.
   * The Server SHOULD compare this hash with the config hash
   * it has for the Agent and if the hashes are different the Server MUST include
   * the remote_config field in the response in the ServerToAgent message.
   *
   * @generated from field: bytes last_remote_config_hash = 1;
   */
  lastRemoteConfigHash: Uint8Array;

  /**
   * @generated from field: opamp.proto.RemoteConfigStatuses status = 2;
   */
  status: RemoteConfigStatuses;

  /**
   * Optional error message if status==FAILED.
   *
   * @generated from field: string error_message = 3;
   */
  errorMessage: string;
};
// ...
export declare type PackageStatuses = Message<"opamp.proto.PackageStatuses"> & {
  /**
   * A map of PackageStatus messages, where the keys are package names.
   * The key MUST match the name field of PackageStatus message.
   *
   * @generated from field: map<string, opamp.proto.PackageStatus> packages = 1;
   */
  packages: { [key: string]: PackageStatus };

  /**
   * The aggregate hash of all packages that this Agent previously received from the
   * Server via PackagesAvailable message.
   *
   * The Server SHOULD compare this hash to the aggregate hash of all packages that
   * it has for this Agent and if the hashes are different the Server SHOULD send
   * an PackagesAvailable message to the Agent.
   *
   * @generated from field: bytes server_provided_all_packages_hash = 2;
   */
  serverProvidedAllPackagesHash: Uint8Array;

  /**
   * This field is set if the Agent encountered an error when processing the
   * PackagesAvailable message and that error is not related to any particular single
   * package.
   * The field must be unset is there were no processing errors.
   *
   * @generated from field: string error_message = 3;
   */
  errorMessage: string;
};
// ...
export declare type CustomCapabilities = Message<"opamp.proto.CustomCapabilities"> & {
  /**
   * A list of custom capabilities that are supported. Each capability is a reverse FQDN
   * with optional version information that uniquely identifies the custom capability
   * and should match a capability specified in a supported CustomMessage.
   * Status: [Development]
   *
   * @generated from field: repeated string capabilities = 1;
   */
  capabilities: string[];
};